HTML encode non-ASCII characters
Since we set `Content-Transfer-Encoding` to `7bit` for HTML messages, we need to ensure that the passed in HTML is 7 bit safe. This is done by replacing any character with higher bits with their HTML entity encoding.

There may be more efficient ways to achieve this, but I've found that working on the string character by character results in the fewest allocations.